requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJWT
django-qrcode
cryptography
Pillow
celery 
redis 
django-celery-beat
gevent
python-dotenv
django-celery-results
django-cors-headers
channels
pycryptodome
psycopg2
django-redis
django-ratelimit
gunicorn
dj-database-url
psycopg2-binary
whitenoise[brotli]
dj-database-url
firebase-admin
prometheus_client
